# packages/proxy/src/openapi.yaml
openapi: 3.1.0
info:
  title: HAAI API
  description: |
    Management REST API and OpenAI-compatible inference endpoints for HAAI.

    **Management API** (`/api/v1/*`) requires admin authentication via session cookie or Bearer token.

    **Inference API** (`/v1/*`) requires an API key in the `Authorization: Bearer` header.
  version: 0.0.1
servers:
  - url: /
    description: Current host
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: haai_session
    bearerAuth:
      type: http
      scheme: bearer
    apiKeyAuth:
      type: http
      scheme: bearer
      description: Client API key (haai-sk-…)
...
paths:
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string, format: password }
      responses:
        "200":
          description: Session established (haai_session cookie set)
        "401":
          description: Invalid credentials
  /api/v1/auth/me:
    get:
      tags: [Auth]
      summary: Current admin user
      security: [{ sessionCookie: [] }, { bearerAuth: [] }]
      responses:
        "200":
          description: Authenticated user profile
        "401":
          description: Not authenticated
  /api/v1/auth/logout:
    post:
      tags: [Auth]
      summary: End admin session
      security: [{ sessionCookie: [] }, { bearerAuth: [] }]
      responses:
        "200":
          description: Logged out
...